types refactor & make sure everything exported as wei
