--- packages/queries/src/queries/shorts/types.ts
@@ -1,8 +1,0 @@
-export type ShortRewardsData = {
-	openInterestUSD: number;
-	distribution: number;
-	periodFinish: number;
-	duration: number;
-	staked: number;
-	rewards: number;
-};
