+++ packages/queries/src/queries/staking/useGetFeePoolDataQuery.ts
@@ -1,16 +1,7 @@
 import { useQuery, UseQueryOptions } from 'react-query';
-import Wei, { wei } from '@synthetixio/wei';
+import { wei } from '@synthetixio/wei';
 import { QueryContext } from '../../context';
-
-export type FeePoolData = {
-	feePeriodDuration: number;
-	startTime: number;
-	feesToDistribute: number;
-	feesClaimed: number;
-	rewardsToDistribute: number;
-	rewardsToDistributeBN: Wei;
-	rewardsClaimed: number;
-};
+import { FeePoolData } from '../../types';
 
 const useGetFeePoolDataQuery = (ctx: QueryContext, period: string, options?: UseQueryOptions<FeePoolData>) => {
 	return useQuery<FeePoolData>(
@@ -18,18 +9,16 @@ const useGetFeePoolDataQuery = (ctx: QueryContext, period: string, options?: Use
 		async () => {
 			const {
 				contracts: { FeePool },
-				utils: { formatEther },
 			} = ctx.snxjs!;
 			const feePeriod = await FeePool.recentFeePeriods(period);
 			const feePeriodDuration = await FeePool.feePeriodDuration();
 			return {
 				feePeriodDuration: Number(feePeriodDuration),
-				startTime: Number(feePeriod.startTime) || 0,
-				feesToDistribute: Number(formatEther(feePeriod.feesToDistribute)) || 0,
-				feesClaimed: Number(formatEther(feePeriod.feesClaimed)) || 0,
-				rewardsToDistribute: Number(formatEther(feePeriod.rewardsToDistribute)) || 0,
-				rewardsToDistributeBN: wei(formatEther(feePeriod.rewardsToDistribute)),
-				rewardsClaimed: Number(formatEther(feePeriod.rewardsClaimed)) || 0,
+				startTime: Number(feePeriod.startTime || 0),
+				feesToDistribute: wei(feePeriod.feesToDistribute || 0),
+				feesClaimed: wei(feePeriod.feesClaimed || 0),
+				rewardsToDistribute: wei(feePeriod.rewardsToDistribute || 0),
+				rewardsClaimed: wei(feePeriod.rewardsClaimed || 0),
 			};
 		},
 		{
